+++ test/html_query_test.exs
@@ -159,19 +159,31 @@ defmodule HtmlQueryTest do
   end
 
   describe "form_fields" do
-    test "returns textareas and inputs that have a `value` attr as a name -> value map" do
+    test "returns selects, textareas, and inputs that have a `value` attr as a name -> value map" do
       """
       <form test-role="test-form">
         <input type="text" name="name" value="alice">
         <input type="number" name="age" value="100">
         <input type="email" name="email" value="">
         <textarea name="about">Alice is 100</textarea>
+        <select name="favorite_color"><option>red</option><option selected>blue</option><option>green</option></select>
         <input type="submit" name="save">
       </form>
       """
       |> Hq.find(test_role: "test-form")
       |> Hq.form_fields()
-      |> assert_eq(%{name: "alice", age: "100", email: "", about: "Alice is 100"})
+      |> assert_eq(%{name: "alice", age: "100", email: "", about: "Alice is 100", favorite_color: "blue"})
+    end
+
+    test "returns empty string for selects with no selection" do
+      """
+      <form test-role="test-form">
+        <select name="favorite_color"><option>red</option><option>blue</option><option>green</option></select>
+      </form>
+      """
+      |> Hq.find(test_role: "test-form")
+      |> Hq.form_fields()
+      |> assert_eq(%{favorite_color: ""})
     end
 
     test "returns a nested map when names are in x[y] format" do
@@ -182,13 +194,14 @@ defmodule HtmlQueryTest do
         <input type="email" name="auth[email]" value="alice@example.com">
         <input type="password" name="auth[password]" value="password123">
         <textarea name="auth[about]">Alice is 100</textarea>
+        <select name="person[favorite_color]"><option selected>red</option><option>blue</option><option>green</option></select>
       </form>
       """
       |> Hq.find("form")
       |> Hq.form_fields()
       |> assert_eq(%{
         _csrf: "_123xyz",
-        person: %{name: "alice"},
+        person: %{name: "alice", favorite_color: "red"},
         auth: %{email: "alice@example.com", password: "password123", about: "Alice is 100"}
       })
     end